+++ src/main/java/com/example/jolvre/auth/login/service/LoginService.java
@@ -3,6 +3,7 @@ package com.example.jolvre.auth.login.service;
 import com.example.jolvre.user.entity.User;
 import com.example.jolvre.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -10,7 +11,9 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class LoginService implements UserDetailsService {
+
     private final UserRepository userRepository;
 
     @Override
